chore: rename src directory
