+++ src/PanasonicCameraFactory.cs
