+++ src/PanasonicCameraPreset.cs
